Update django-browserid and replace submit button when logged out.
If a user is logged out and clicks the submit button, the site currently
shows them a page in the frame that auto-triggers Persona login, causing
a popup warning. This patch fixes this by instead replacing the submit
button with a login button when the user isn't logged in, and uses an 
anchor in the URL to trigger the submission modal after login is 
finished.