Temporarily redirect invalid locales to the homepage.
Funfactory's LocaleURLMiddleware will tack en-US onto any URL without an
active locale, such as /fr -> /en-US/fr/. This causes 404s when it tries
to match on the url /fr.

This is a temporary fix that redirects any path that starts with a
locale-like string to the homepage. This should catch any redirects with
locales in the path and just push users towards the homepage. We'll 
remove this once the locales we want are back in.