Bug 886500: Add autocomplete dropdown.
Adds a JS-powered widget to the search page that shows suggested
completions for a user's search term. It's not super-smart; in the
background it performs the search and returns the first result found,
which isn't weighted by a score or anything, so it's mostly a shot in
the dark. But it's not terrible if you're searching for a semi-unique
string; the primary use-case here is people searching for videos by
their friends.

The design of the frontend bits are in one of craigcook's pull
requests, and I'll be integrating it later. The JavaScript is
functionally complete, although it will be modified to fit whatever
frontend craigcook came up with.

I decided to try a different approach to the JavaScript bits by
separating the widget logic with the DOM-handling code. It's a little
verbose in this case, but I'm actually quite happy with the result.