Fix bug 904281: Add more ways to see stats about votes to the admin.
Adds a few niceties to the admin interface to make it easier to find
information about how many votes there are, who has voted, what they
have voted on, etc.

There's no tests because admin stuff is in general hard to test and
the set of users for this stuff is small and limited to trusted 
individuals. The only testable thing here is the admin filter, but it's
trivial and doesn't handle any edge cases.